update-exception-delete-book

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -61,11 +61,10 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
-        // try {
-        //     User::where('id', $id)->firstOrFail()->delete();
-        // } catch (ModelCouldNotDeletedException $exception) {
-        //     return (['error' => 'U cant delete it']);
-        // }
+        if ($id = 1) {
+            return redirect()->back()->with("error", "You can't delete this !! It's Default");
+        }
+
         Category::where('id', $id)->firstOrFail()->delete();
 
         return redirect()->back();
